Moved the AppBar to separate components. Separated the navigation routes and the settings. I do not know should they be left or right.
